move files from actions to controller
