+++ docker/go/src/controller/make-buy-order.go
@@ -1,4 +1,4 @@
-package actions
+package controller
 
 import (
 	"model"
