+++ docker/go/src/controller/make-sell-order.go
@@ -1,4 +1,4 @@
-package actions
+package controller
 
 import (
 	"model"
