+++ docker/go/src/controller/watch-buy-execution.go
@@ -1,4 +1,4 @@
-package actions
+package controller
 
 import (
 	"model"
