+++ docker/go/src/job/auto-trading-job.go
@@ -1,10 +1,9 @@
 package job
 
 import (
-	"actions"
 	"fmt"
 	"model"
-
+	"controller"
 	"github.com/carlescere/scheduler"
 )
 
@@ -13,16 +12,16 @@ func AutoTradingJob() {
 		tradeMode := model.CheckTradeMode()
 		if tradeMode == 0 {
 			fmt.Println("Mode is 0,now func MakeBuyOrder is going to be executed.")
-			actions.MakeBuyOrder()
+			controller.MakeBuyOrder()
 		} else if tradeMode == 1 {
 			fmt.Println("Mode is 1,now func WatchBuyExecution is going to be executed.")
-			actions.WatchBuyExecution()
+			controller.WatchBuyExecution()
 		} else if tradeMode == 2 {
 			fmt.Println("Mode is 2,now func makeSellOrder is going to be executed.")
-			actions.MakeSellOrder()
+			controller.MakeSellOrder()
 		} else if tradeMode == 3 {
 			fmt.Println("Mode is 3,now func makeSellOrder is going to be executed.")
-			actions.WatchSellExecution()
+			controller.WatchSellExecution()
 		}
 	}
 	scheduler.Every(10).Seconds().Run(job)
